refactor: remove python podman client
Unfortunately, the official Python Podman client is not available in
standard RHEL8 and RHEL9 packages for Python 3.12. So, we are
completely excising its use from quipucordsctl, and this change
replaces any uses of the client with bare subprocess calls to the
podman CLI.

While replacing that logic, this commit also cleans up some
inconsistent log messages, exception handling, and function returns.

Relates to JIRA: DISCOVERY-1158