chore: remove .python-version file
`uv` will try to use Python defined inside `.python-version` file, and
refuse to run if it can't. If people want to force a specific Python
version for the project, that's fine, but there's no reason to force
everyone down the same path - people should be free to use any version
they want, as long as it matches `pyproject.toml` specification.